refactor: move generate_invalid_id and generate_openssh_pkey to tests.factories so actual test modules aren't importing from each other
