+++ quipucords/tests/scanner/network/test_network_inspect.py
@@ -23,7 +23,7 @@ from scanner.network.exceptions import NetworkCancelException, NetworkPauseExcep
 from scanner.network.inspect import construct_inventory
 from scanner.network.inspect_callback import InspectCallback
 from scanner.network.utils import delete_ssh_keyfiles, is_gen_ssh_keyfile
-from tests.api.credential.test_credential import generate_openssh_pkey
+from tests.factories import generate_openssh_pkey
 from tests.scanner.test_util import create_scan_job
 
 ANSIBLE_FACTS = "ansible_facts"
